:shirt: make comparison in tests in tests phase
Before of this commit, check_flow() was done at the end of test phase
(teardown). Because of this, all comparison were done in this phase, so
errors were shown as "test suite errors", not test errors. From now on,
errors will be shown as actual cmocka errors.

Also, test 0005 had to be changed because of not to check actual output,
but only check if they had produced output or not.